+++ src/Benzene.Xml/Settings.cs
@@ -1,7 +1,0 @@
-﻿namespace Benzene.Xml;
-
-public static class Settings
-{
-    public static string ContentTypeKey { get; set; } = Constants.ContentTypeHeader;
-    public static string ContentTypeValue { get; set; } = Constants.XmlContentType;
-}